requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
psycopg2-binary

jinja2
python-multipart
python-dotenv

bcrypt==4.0.1
python-jose[cryptography]